Add Share btn to ImgUtils component
The ShareBtn takes in a callback function, rather than canvas like DownloadBtn does.
The reasoning is that, after the image is uploaded then it has to return a URL to the user.
At the moment I plan to show it as a green confirmation line above the other utility buttons,
so it would be best that ImgUtils handles that rather than ShareBtn itself.