Add auth0 clientID/domain explicitly
Initially I had it as part of the app-only .env however
when deploying to heroku, the current docker config
does not pull in `app/.env`. For the sake of others
being able to use my code in general on github I left in
the spots for people to fill it in with their own info.
However, with working with my peers this has proven
ineffective. These two are safe to publically release,
but in final iterations I want to investigate the docker
issue further.