User upload also generates a url with 24-hr timer
I need to set access-control on that url, from the S3 side such that it can only be fetched by my domain, no one else. This can be used to generate a scribble thumbnail on the history page.

This is a bit excessive in requests, and something like a signed cookie by Cloudfront might be more efficient. However, in the interest of time, this will have to do.